requirements.txt
amqp==5.2.0
aniso8601==9.0.1
async-timeout==4.0.3
attrs==23.2.0
billiard==4.2.0
blinker==1.7.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
dnspython==2.5.0
Flask==3.0.2
flask-restx==1.3.0
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kombu==5.3.5
MarkupSafe==2.1.5
pillow==10.2.0
prompt-toolkit==3.0.43
pymongo==4.6.1
python-dateutil==2.8.2
pytz==2024.1
redis==5.0.1
referencing==0.33.0
rpds-py==0.18.0
six==1.16.0
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.1